move buildDistanceInWordsLocale on a specify file
