+++ packages/components/src/VirtualizedList/CellDatetime/CellDatetime.test.js
@@ -2,101 +2,13 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { distanceInWordsToNow, format } from 'date-fns';
 
-import {
-	computeValue,
-	CellDatetimeComponent,
-	buildDistanceInWordsLocale,
-} from './CellDatetime.component';
+import { computeValue, CellDatetimeComponent } from './CellDatetime.component';
 import getDefaultT from '../../translate';
 
 jest.mock('date-fns', () => ({
 	format: jest.fn(() => '2016-09-22 09:00:00'),
 	distanceInWordsToNow: jest.fn(() => 'about 1 month ago'),
 }));
-
-describe('wrapT', () => {
-	it('should return the formatDistance with ', () => {
-		const distanceInWords = buildDistanceInWordsLocale(getDefaultT());
-		expect(
-			distanceInWords.localize('lessThanXSeconds', 5, {
-				addSuffix: true,
-			}),
-		).toBe('less than 5 seconds ago');
-		expect(
-			distanceInWords.localize('xSeconds', 5, {
-				addSuffix: true,
-			}),
-		).toBe('5 seconds ago');
-		expect(
-			distanceInWords.localize('halfAMinute', 5, {
-				addSuffix: true,
-			}),
-		).toBe('half a minute ago');
-		expect(
-			distanceInWords.localize('lessThanXMinutes', 5, {
-				addSuffix: true,
-			}),
-		).toBe('less than 5 minutes ago');
-		expect(
-			distanceInWords.localize('xMinutes', 5, {
-				addSuffix: true,
-			}),
-		).toBe('5 minutes ago');
-		expect(
-			distanceInWords.localize('aboutXHours', 5, {
-				addSuffix: true,
-			}),
-		).toBe('about 5 hours ago');
-		expect(
-			distanceInWords.localize('xHours', 5, {
-				addSuffix: true,
-			}),
-		).toBe('5 hours ago');
-		expect(
-			distanceInWords.localize('xDays', 5, {
-				addSuffix: true,
-			}),
-		).toBe('5 days ago');
-		expect(
-			distanceInWords.localize('aboutXMonths', 5, {
-				addSuffix: true,
-			}),
-		).toBe('about 5 months ago');
-		expect(
-			distanceInWords.localize('xMonths', 5, {
-				addSuffix: true,
-			}),
-		).toBe('5 months ago');
-		expect(
-			distanceInWords.localize('aboutXYears', 5, {
-				addSuffix: true,
-			}),
-		).toBe('about 5 years ago');
-		expect(
-			distanceInWords.localize('xYears', 5, {
-				addSuffix: true,
-			}),
-		).toBe('5 years ago');
-		expect(
-			distanceInWords.localize('overXYears', 5, {
-				addSuffix: true,
-			}),
-		).toBe('over 5 years ago');
-		expect(
-			distanceInWords.localize('almostXYears', 5, {
-				addSuffix: true,
-			}),
-		).toBe('almost 5 years ago');
-		expect(
-			distanceInWords.localize('lessThanXSeconds', 5, {
-				addSuffix: true,
-				comparison: 1,
-			}),
-		).toBe('in less than 5 seconds');
-
-		expect(distanceInWords.localize('lessThanXSeconds', 5)).toBe('less than 5 seconds');
-	});
-});
 
 describe('CellDatetime', () => {
 	it('should render CellDatetime', () => {
